[SR-3289] FE schedule algorithm for backend disk balance (#119)
1. On the basis of the original cluster tablet balance, add backend disk balance schedule algorithm.
2. Implement backend local migration schedule process.
3. Improve the storage medium migration process.
4. Add storage medium, tablet num in BackendProcNode.

+++ fe/fe-core/src/main/java/com/starrocks/metric/MetricRepo.java
@@ -423,7 +423,7 @@ public final class MetricRepo {
                     if (!Catalog.getCurrentCatalog().isMaster()) {
                         return 0L;
                     }
-                    return (long) invertedIndex.getTabletNumByBackendId(beId);
+                    return invertedIndex.getTabletNumByBackendId(beId);
                 }
             };
             tabletNum.addLabel(new MetricLabel("backend", be.getHost() + ":" + be.getHeartbeatPort()));
